exp: save: Remove metrics.
Move tests to separate file.
